tired of launchers not working because there isn't a .desktop file
we KNOW when there is an exec that we can run it; for the broken apps out there
that do not have a .desktop file, are not in the fix-up lists AND can not be run
from their executable .. screw 'em. i'd rather have the real world work even if
that means edge case applications written by idiots fail.