MPRIS2: assume the position is 0 when stopped
MPRIS2 does not require the Seeked() signal to be emitted when stopping
a track.  It does, however, require it to be emitted if the position is
not 0 when going from Stopped to Playing (or Paused).

So we should assume the position is 0 when Stopped.

This fixes an issue I noticed with gmusicbrowser, where stopping a track
and starting it again would show the wrong time in the Now Playing
widget.

Backport of dc7074f9f503aac54007952d7429d8f1eb6a80c2

FIXED-IN: 4.9.1